Create README.md
We are creating a GitHub account. 
In this account we create a private GitHub repository for every project. 
In project 1's repository we create a project and link it with the repository.
In the project we make table and board views.
With this views we track our progress and plan out our tasks to be done.
We label the tasks and give them milestones.
We then configure charts from the views.